training/requirements.txt
torch>=2.0
torchvision>=0.15
timm>=1.0
opencv-python-headless>=4.8
Pillow>=10.0
tqdm
